requirements.txt
streamlit>=1.32.0
supabase>=2.3.6
google-generativeai>=0.4.0
pandas>=2.2.1
scikit-learn>=1.4.1.post1
plotly>=5.19.0
requests>=2.31.0